Moves salt generation object to crypto package
